Fix & Organize Tests
