+++ Thirdweb.Tests/BaseTests.cs
@@ -16,9 +16,11 @@ public class BaseTests
         _secretKey = Environment.GetEnvironmentVariable("THIRDWEB_SECRET_KEY");
         _clientIdBundleIdOnly = Environment.GetEnvironmentVariable("THIRDWEB_CLIENT_ID_BUNDLE_ID_ONLY");
         _bundleIdBundleIdOnly = Environment.GetEnvironmentVariable("THIRDWEB_BUNDLE_ID_BUNDLE_ID_ONLY");
+
+        _output.WriteLine($"Initialized {GetType().Name}");
     }
 
-    [Fact]
+    [Fact(Timeout = 120000)]
     public void DotEnvTest()
     {
         Assert.NotNull(_secretKey);
