+++ Thirdweb.Tests/Thirdweb.Tests.csproj
@@ -39,5 +39,11 @@
     </None>
   </ItemGroup>
 
+  <ItemGroup>
+    <None Update="xunit.runner.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+
 
 </Project>
